Score system and end of game implemented
End of the round is met if by the end of the missile fire cities
are still standing. The end of the game is met if every city is
destroyed.

Missiles now split if they reach halfway down the terminal height. Each
missile fired by the AI is capable of splitting into another
missile once per flight.